Refactor HTTP package structure and update imports
Moved HTTP-related enums and models into 'enumerator', 'model', and 'handler' subpackages for better organization. Updated all relevant imports and references throughout the codebase to reflect the new structure. No functional changes were made.
