--- src/main/java/nl/radiantrealm/library/http/HttpHandler.java
@@ -1,5 +1,0 @@
-package nl.radiantrealm.library.http;
-
-public interface HttpHandler {
-    void handle(HttpRequest request) throws Exception;
-}
