+++ src/main/java/nl/radiantrealm/library/http/enumerator/MediaType.java
@@ -1,4 +1,4 @@
-package nl.radiantrealm.library.http;
+package nl.radiantrealm.library.http.enumerator;
 
 public enum MediaType {
     TEXT("text/plain"),
