+++ src/main/java/nl/radiantrealm/library/http/enumerator/StatusCode.java
@@ -1,9 +1,7 @@
-package nl.radiantrealm.library.http;
+package nl.radiantrealm.library.http.enumerator;
 
 import com.google.gson.JsonObject;
 import nl.radiantrealm.library.utils.DataObject;
-
-import javax.security.auth.kerberos.KeyTab;
 
 public enum StatusCode implements DataObject {
     CONTINUE(100, null),
