+++ src/main/java/nl/radiantrealm/library/http/handler/HttpHandler.java
@@ -1,0 +1,7 @@
+package nl.radiantrealm.library.http.handler;
+
+import nl.radiantrealm.library.http.model.HttpRequest;
+
+public interface HttpHandler {
+    void handle(HttpRequest request) throws Exception;
+}
